+++ saleor/graphql/product/tests/test_product.py
@@ -2384,6 +2384,7 @@ CREATE_PRODUCT_MUTATION = """
                                     slug
                                     name
                                     reference
+                                    richText
                                     file {
                                         url
                                         contentType
@@ -2533,6 +2534,76 @@ def test_create_product_description_plaintext(
 
     product = Product.objects.all().first()
     assert product.description_plaintext == description
+
+
+def test_create_product_with_rich_text_attribute(
+    staff_api_client,
+    product_type,
+    category,
+    rich_text_attribute,
+    color_attribute,
+    permission_manage_products,
+    product,
+):
+    query = CREATE_PRODUCT_MUTATION
+
+    product_type_id = graphene.Node.to_global_id("ProductType", product_type.pk)
+    category_id = graphene.Node.to_global_id("Category", category.pk)
+    product_name = "test name"
+    product_slug = "product-test-slug"
+
+    # Add second attribute
+    product_type.product_attributes.add(rich_text_attribute)
+    rich_text_attribute_id = graphene.Node.to_global_id(
+        "Attribute", rich_text_attribute.id
+    )
+    rich_text = json.dumps(dummy_editorjs("test product" * 5))
+
+    # test creating root product
+    variables = {
+        "input": {
+            "productType": product_type_id,
+            "category": category_id,
+            "name": product_name,
+            "slug": product_slug,
+            "attributes": [
+                {
+                    "id": rich_text_attribute_id,
+                    "richText": rich_text,
+                }
+            ],
+        }
+    }
+
+    response = staff_api_client.post_graphql(
+        query, variables, permissions=[permission_manage_products]
+    )
+    content = get_graphql_content(response)
+    data = content["data"]["productCreate"]
+    assert data["productErrors"] == []
+    assert data["product"]["name"] == product_name
+    assert data["product"]["slug"] == product_slug
+    assert data["product"]["productType"]["name"] == product_type.name
+    assert data["product"]["category"]["name"] == category.name
+    _, product_id = graphene.Node.from_global_id(data["product"]["id"])
+
+    expected_attributes_data = [
+        {"attribute": {"slug": "color"}, "values": []},
+        {
+            "attribute": {"slug": "text"},
+            "values": [
+                {
+                    "slug": f"{product_id}_{rich_text_attribute.id}",
+                    "name": "test producttest producttest producttest productt…",
+                    "reference": None,
+                    "richText": rich_text,
+                    "file": None,
+                }
+            ],
+        },
+    ]
+    for attr_data in data["product"]["attributes"]:
+        assert attr_data in expected_attributes_data
 
 
 SEARCH_PRODUCTS_QUERY = """
@@ -2776,6 +2847,7 @@ def test_create_product_with_file_attribute(
                     "slug": f"{existing_value.slug}-2",
                     "file": {"url": existing_value.file_url, "contentType": None},
                     "reference": None,
+                    "richText": None,
                 }
             ],
         },
@@ -2843,6 +2915,7 @@ def test_create_product_with_page_reference_attribute(
                     "slug": f"{product_id}_{page.id}",
                     "name": page.title,
                     "file": None,
+                    "richText": None,
                     "reference": reference,
                 }
             ],
@@ -2911,6 +2984,7 @@ def test_create_product_with_product_reference_attribute(
                     "slug": f"{product_id}_{product.id}",
                     "name": product.name,
                     "file": None,
+                    "richText": None,
                     "reference": reference,
                 }
             ],
@@ -2979,6 +3053,7 @@ def test_create_product_with_product_reference_attribute_values_saved_in_order(
             "slug": f"{product_id}_{product.id}",
             "name": product.name,
             "file": None,
+            "richText": None,
             "reference": reference,
         }
         for product, reference in zip(reference_instances, reference_ids)
@@ -3050,6 +3125,7 @@ def test_create_product_with_file_attribute_new_attribute_value(
                     "name": non_existing_value,
                     "slug": slugify(non_existing_value, allow_unicode=True),
                     "reference": None,
+                    "richText": None,
                     "file": {
                         "url": "http://testserver/media/" + non_existing_value,
                         "contentType": None,
@@ -5154,7 +5230,13 @@ DELETE_PRODUCT_MUTATION = """
 """
 
 
-def test_delete_product(staff_api_client, product, permission_manage_products):
+@patch("saleor.order.tasks.recalculate_orders_task.delay")
+def test_delete_product(
+    mocked_recalculate_orders_task,
+    staff_api_client,
+    product,
+    permission_manage_products,
+):
     query = DELETE_PRODUCT_MUTATION
     node_id = graphene.Node.to_global_id("Product", product.id)
     variables = {"id": node_id}
@@ -5167,10 +5249,13 @@ def test_delete_product(staff_api_client, product, permission_manage_products):
     with pytest.raises(product._meta.model.DoesNotExist):
         product.refresh_from_db()
     assert node_id == data["product"]["id"]
+    mocked_recalculate_orders_task.assert_called_once_with(set())
 
 
 @patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@patch("saleor.order.tasks.recalculate_orders_task.delay")
 def test_delete_product_trigger_webhook(
+    mocked_recalculate_orders_task,
     mocked_webhook_trigger,
     staff_api_client,
     product,
@@ -5198,9 +5283,12 @@ def test_delete_product_trigger_webhook(
     mocked_webhook_trigger.assert_called_once_with(
         WebhookEventType.PRODUCT_DELETED, expected_data
     )
+    mocked_recalculate_orders_task.assert_called_once_with(set())
 
 
+@patch("saleor.order.tasks.recalculate_orders_task.delay")
 def test_delete_product_variant_in_draft_order(
+    mocked_recalculate_orders_task,
     staff_api_client,
     product_with_two_variants,
     permission_manage_products,
@@ -5266,6 +5354,7 @@ def test_delete_product_variant_in_draft_order(
     assert not OrderLine.objects.filter(pk__in=draft_order_lines_pks).exists()
 
     assert OrderLine.objects.filter(pk__in=not_draft_order_lines_pks).exists()
+    mocked_recalculate_orders_task.assert_called_once_with({draft_order.id})
 
 
 def test_product_type(user_api_client, product_type, channel_USD):
@@ -5485,6 +5574,7 @@ PRODUCT_TYPE_CREATE_MUTATION = """
                     name
                     values {
                         name
+                        richText
                     }
                 }
             }
@@ -5561,6 +5651,62 @@ def test_product_type_create_mutation(
     new_instance = ProductType.objects.latest("pk")
     tax_code = manager.get_tax_code_from_object_meta(new_instance).code
     assert tax_code == "wine"
+
+
+def test_create_product_type_with_rich_text_attribute(
+    staff_api_client,
+    product_type,
+    permission_manage_product_types_and_attributes,
+    rich_text_attribute,
+):
+    query = PRODUCT_TYPE_CREATE_MUTATION
+    product_type_name = "test type"
+    slug = "test-type"
+
+    product_type.product_attributes.add(rich_text_attribute)
+    product_attributes_ids = [
+        graphene.Node.to_global_id("Attribute", attr.id)
+        for attr in product_type.product_attributes.all()
+    ]
+
+    variables = {
+        "name": product_type_name,
+        "slug": slug,
+        "productAttributes": product_attributes_ids,
+    }
+
+    response = staff_api_client.post_graphql(
+        query, variables, permissions=[permission_manage_product_types_and_attributes]
+    )
+    content = get_graphql_content(response)
+    data = content["data"]["productTypeCreate"]["productType"]
+    errors = content["data"]["productTypeCreate"]["productErrors"]
+
+    assert not errors
+    assert data["name"] == product_type_name
+    assert data["slug"] == slug
+    expected_attributes = [
+        {
+            "name": "Color",
+            "values": [
+                {"name": "Red", "richText": None},
+                {"name": "Blue", "richText": None},
+            ],
+        },
+        {
+            "name": "Text",
+            "values": [
+                {
+                    "name": "Rich text attribute content.",
+                    "richText": json.dumps(
+                        rich_text_attribute.values.first().rich_text
+                    ),
+                }
+            ],
+        },
+    ]
+    for attribute in data["productAttributes"]:
+        assert attribute in expected_attributes
 
 
 @pytest.mark.parametrize(